refactored quickgen to include quickbrick
