--- bin/quickbrick
@@ -1,16 +1,0 @@
-#!/usr/bin/env python
-#
-# See top-level LICENSE.rst file for Copyright information
-#
-# -*- coding: utf-8 -*-
-
-"""Quickly generate DESI brick files.
-
-TODO (@moustakas): Use the correct fiberloss file for BGS objects.
-"""
-
-import desisim.scripts.quickbrick as quickbrick
-
-if __name__ == '__main__':
-    args = quickbrick.parse()
-    quickbrick.main(args)
