+++ py/desisim/test/test_quickgen.py
@@ -10,7 +10,7 @@ import numpy as np
 import desispec.io
 from desisim import io
 from desisim import obs
-import desisim.scripts.quickgen
+from desisim.scripts import quickgen
 from desispec.log import get_logger
 log = get_logger()
 
@@ -133,14 +133,14 @@ class TestQuickgen(unittest.TestCase):
         fibermap = desispec.io.findfile('fibermap', self.night, self.expid)
         opts = ['--simspec', simspec, '--fibermap', fibermap]
         opts += ['--spectrograph', '3', '--config', 'desi']
-        args = desisim.scripts.quickgen.parse(opts)
+        args = quickgen.parse(opts)
         self.assertEqual(args.simspec, simspec)
         self.assertEqual(args.fibermap, fibermap)
         self.assertEqual(args.spectrograph, 3)
         self.assertEqual(args.config, 'desi')
 
         with self.assertRaises(ValueError):
-            desisim.scripts.quickgen.parse([])
+            quickgen.parse([])
 
     def test_expand_args(self):
         night = self.night
@@ -150,17 +150,18 @@ class TestQuickgen(unittest.TestCase):
         simspec = io.findfile('simspec', self.night, self.expid)
         fibermap = desispec.io.findfile('fibermap', self.night, self.expid)
         opts = ['--simspec', simspec, '--fibermap', fibermap]
-        args = desisim.scripts.quickgen.parse(opts)
+        args = quickgen.parse(opts)
         self.assertEqual(args.simspec, simspec)
         self.assertEqual(args.fibermap, fibermap)
 
     @unittest.skipUnless(desi_root_available, '$DESI_ROOT not set')
+    #- Run basic test of quickgen and check its outputs using simspec as input
     ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
     def test_quickgen(self):
         night = self.night
         expid = self.expid
         camera = 'r0'
-        # flavors = ['flat','dark','gray','bright','bgs','mws','elg','lrg','qso']#,'arc']
+        # flavors = ['flat','dark','gray','bright','bgs','mws','elg','lrg','qso','arc']
         flavors = ['arc', 'flat', 'dark', 'bright']
         for i in range(len(flavors)):
             flavor = flavors[i]
@@ -172,12 +173,11 @@ class TestQuickgen(unittest.TestCase):
             #- run quickgen
             simspec = io.findfile('simspec', self.night, self.expid)
             fibermap = desispec.io.findfile('fibermap', self.night, self.expid)
-            opts = ['--simspec', simspec, '--fibermap', fibermap]
-            log.debug('testing quickgen({})'.format(opts))
 
             if flavor == 'flat':
                 try:
-                    desisim.scripts.quickgen.main(opts)
+                    cmd = "quickgen --simspec {} --fibermap {}".format(simspec,fibermap)
+                    quickgen.main(quickgen.parse(cmd.split()[1:]))
                 except SystemExit:
                     pass
 
@@ -185,12 +185,10 @@ class TestQuickgen(unittest.TestCase):
                 fiberflatfile = desispec.io.findfile('fiberflat', night, expid, camera)
                 self.assertTrue(os.path.exists(fiberflatfile))
 
-                #- cleanup flat outputs
-                os.remove(fiberflatfile)
-
             elif flavor == 'arc':
                 try:
-                    desisim.scripts.quickgen.main(opts)
+                    cmd = "quickgen --simspec {} --fibermap {}".format(simspec,fibermap)
+                    quickgen.main(quickgen.parse(cmd.split()[1:]))
                 except SystemExit:
                     pass
 
@@ -198,11 +196,9 @@ class TestQuickgen(unittest.TestCase):
                 framefile = desispec.io.findfile('frame', night, expid, camera)
                 self.assertTrue(os.path.exists(framefile))
 
-                #- cleanup arc outputs
-                os.remove(framefile)
-
             else:
-                desisim.scripts.quickgen.main(opts)
+                cmd = "quickgen --simspec {} --fibermap {}".format(simspec,fibermap)
+                quickgen.main(quickgen.parse(cmd.split()[1:]))
     
                 #- verify outputs
                 framefile = desispec.io.findfile('frame', night, expid, camera)
@@ -213,46 +209,106 @@ class TestQuickgen(unittest.TestCase):
                 self.assertTrue(os.path.exists(skyfile))
                 fluxcalibfile = desispec.io.findfile('calib', night, expid, camera)
                 self.assertTrue(os.path.exists(fluxcalibfile))
-    
-                #- cleanup outputs
-                os.remove(framefile)
-                os.remove(cframefile)
-                os.remove(skyfile)
-                os.remove(fluxcalibfile)
 
+    #- Run basic test of quickgen and check its outputs for bricks
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickbrick/specsim test on Travis')
+    def test_quickgen_brick(self):
+        brickname = 'test1'
+        nspec = 3
+        cmd = "quickgen -b {} --objtype ELG -n {} --outdir {}".format(brickname, nspec, self.testDir)
+        args = quickgen.parse(cmd.split()[1:])
+        results = quickgen.main(args)
+        #- Do the bricks exist?
+        for channel in ['z', 'r', 'b']:
+            brickfile = '{}/brick-{}-{}.fits'.format(self.testDir, channel, brickname)
+            self.assertTrue(os.path.exists(brickfile))
+
+        #- Check one of them for correct dimensionality of contents
+        brickfile = '{}/brick-b-{}.fits'.format(self.testDir, brickname)
+        with fits.open(brickfile) as fx:
+            self.assertEqual(fx['FLUX'].shape[0], nspec)
+            self.assertEqual(fx['FLUX'].shape, fx['IVAR'].shape)
+            self.assertEqual(fx['FLUX'].shape[1], fx['WAVELENGTH'].shape[0])
+            self.assertEqual(fx['FLUX'].shape[0], fx['RESOLUTION'].shape[0])
+            self.assertEqual(fx['FLUX'].shape[1], fx['RESOLUTION'].shape[2])
+            self.assertEqual(len(fx['FIBERMAP'].data), nspec)
+            self.assertEqual(fx['_TRUEFLUX'].shape, fx['FLUX'].shape)
+            self.assertTrue(np.all(fx['_TRUTH'].data['OBJTYPE'] == 'ELG'))
+
+        #- Make sure that desispec I/O functions can read it
+        frame = desispec.io.read_frame(brickfile)
+        self.assertEqual(frame.flux.shape[0], nspec)
+        brick = desispec.io.Brick(brickfile)
+        self.assertEqual(brick.get_num_spectra(), nspec)
+        brick.close()
+
+
+    #- Test quickgen with a bunch of options for bricks
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickbrick/specsim test on Travis')
+    def test_quickgen_options_brick(self):
+        brickname = 'test2'
+        nspec = 5
+        cmd = "quickgen -b {} --objtype BGS -n {} --outdir {}".format(brickname, nspec, self.testDir)
+        cmd = cmd + " --airmass 1.5 --verbose --zrange-bgs 0.1 0.2"
+        cmd = cmd + " --moon-phase 0.1 --moon-angle 30 --moon-zenith 20"
+        args = quickgen.parse(cmd.split()[1:])
+        results = quickgen.main(args)
+        brickfile = '{}/brick-b-{}.fits'.format(self.testDir, brickname)
+        truth = fits.getdata(brickfile, '_TRUTH')
+        z = truth['TRUEZ']
+        self.assertTrue(np.all((0.1 <= z) & (z <= 0.2)))
+
+        #- Basic empirical sanity check of median(error) vs. median(flux)
+        #- These coefficients are taken from 50 quickbrick BGS targets
+        #- The main point is to trigger if something significantly changed
+        #- in a new version, not to validate that this is actually the right
+        #- error vs. flux model.
+        coeff = dict()
+        coeff['b'] = [-2.211e-07, 4.490e-05, 1.525e-02, 7.282e+00]
+        coeff['r'] = [1.128e-07, -6.250e-05, 2.423e-02, 3.454e+00]
+        coeff['z'] = [1.394e-07, -7.618e-05, 2.476e-02, 1.953e+00]
+        for channel in ['b', 'r', 'z']:
+            brickfile = '{}/brick-{}-{}.fits'.format(self.testDir, channel, brickname)
+            brick = desispec.io.read_frame(brickfile)
+            medflux = np.median(brick.flux, axis=1)
+            mederr = np.median(1/np.sqrt(brick.ivar), axis=1)
+            std = np.std(mederr - np.polyval(coeff[channel], medflux))
+            self.assertLess(std, 0.2)
+
+    #- Ensure that using --seed results in reproducible spectra using simspec as input
     ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
-    # test to see if same seed yields same spectrum
-    def test_quickgen_seed(self):
-
-        CFRAME100_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000100','cframe-r0-00000100.fits')
-        CFRAME101_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000101','cframe-r0-00000101.fits')
-        CFRAME102_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000102','cframe-r0-00000102.fits')
+    def test_quickgen_seed_simspec(self):
+        night=self.night
+        camera = 'r0'
+        expid0 = 100
+        expid1 = 101
+        expid2 = 102
 
         # generate exposures seed 1 & 2
-        night=self.night
-        obs.new_exposure('dark',night=night,expid=100,nspec=1,seed=1)
-        simspec0 = io.findfile('simspec', night, 100)
-        fibermap0 = desispec.io.findfile('fibermap', night, 100)
-        opts0 = ['--simspec', simspec0, '--fibermap', fibermap0]
-
-        obs.new_exposure('dark',night=night,expid=101,nspec=1,seed=1)
-        simspec1 = io.findfile('simspec', night, 101)
-        fibermap1 = desispec.io.findfile('fibermap', night, 101)
-        opts1 = ['--simspec', simspec1, '--fibermap', fibermap1]
-
-        obs.new_exposure('dark',night=night,expid=102,nspec=1,seed=2)
-        simspec2 = io.findfile('simspec', night, 102)
-        fibermap2 = desispec.io.findfile('fibermap', night, 102)
-        opts2 = ['--simspec', simspec2, '--fibermap', fibermap2]
+        obs.new_exposure('dark',night=night,expid=expid0,nspec=1,seed=1)
+        simspec0 = io.findfile('simspec', night, expid0)
+        fibermap0 = desispec.io.findfile('fibermap', night, expid0)
+        obs.new_exposure('dark',night=night,expid=expid1,nspec=1,seed=1)
+        simspec1 = io.findfile('simspec', night, expid1)
+        fibermap1 = desispec.io.findfile('fibermap', night, expid1)
+        obs.new_exposure('dark',night=night,expid=expid2,nspec=1,seed=2)
+        simspec2 = io.findfile('simspec', night, expid2)
+        fibermap2 = desispec.io.findfile('fibermap', night, expid2)
 
         # generate quickgen output for each exposure
-        desisim.scripts.quickgen.main(opts0)
-        desisim.scripts.quickgen.main(opts1)
-        desisim.scripts.quickgen.main(opts2)
+        cmd = "quickgen --simspec {} --fibermap {} --seed 1".format(simspec0,fibermap0)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --simspec {} --fibermap {} --seed 1".format(simspec1,fibermap1)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --simspec {} --fibermap {} --seed 2".format(simspec2,fibermap2)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        cf0=desispec.io.read_frame(CFRAME100_PATH)
-        cf1=desispec.io.read_frame(CFRAME101_PATH)
-        cf2=desispec.io.read_frame(CFRAME102_PATH)
+        cframe0=desispec.io.findfile("cframe",night,expid0,camera)
+        cframe1=desispec.io.findfile("cframe",night,expid1,camera)
+        cframe2=desispec.io.findfile("cframe",night,expid2,camera)
+        cf0=desispec.io.read_frame(cframe0)
+        cf1=desispec.io.read_frame(cframe1)
+        cf2=desispec.io.read_frame(cframe2)
 
         self.assertTrue(np.all(cf0.flux == cf1.flux))   #- same seed
         self.assertTrue(np.all(cf0.ivar == cf1.ivar))
@@ -266,140 +322,157 @@ class TestQuickgen(unittest.TestCase):
         self.assertTrue(s0[13].data['REDSHIFT'][0] == s1[13].data['REDSHIFT'][0])
         self.assertTrue(s0[13].data['REDSHIFT'][0] != s2[13].data['REDSHIFT'][0])
 
-    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
-    # test to see if increased airmass yields smaller ivar
-    def test_quickgen_airmass(self):
+    #- Ensure that using --seed results in reproducible spectra for bricks
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickbrick/specsim test on Travis')
+    def test_quickgen_seed_brick(self):
+        nspec = 2
+        cmd = "quickgen --seed 1 -b test1a --objtype BRIGHT_MIX -n {} --outdir {}".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --seed 1 -b test1b --objtype BRIGHT_MIX -n {} --outdir {}".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --seed 2 -b test2 --objtype BRIGHT_MIX -n {} --outdir {}".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        CFRAME100_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000100','cframe-r0-00000100.fits')
-        CFRAME101_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000101','cframe-r0-00000101.fits')
+        f1a = desispec.io.read_frame('{}/brick-b-test1a.fits'.format(self.testDir))
+        f1b = desispec.io.read_frame('{}/brick-b-test1b.fits'.format(self.testDir))
+        f2 = desispec.io.read_frame('{}/brick-b-test2.fits'.format(self.testDir))
+        self.assertTrue(np.all(f1a.flux == f1b.flux))  #- same seed
+        self.assertTrue(np.all(f1a.ivar == f1b.ivar))
+        self.assertTrue(np.any(f1a.flux != f2.flux))   #- different seed
+
+        t1a = fits.getdata('{}/brick-b-test1a.fits'.format(self.testDir), '_TRUTH')
+        t1b = fits.getdata('{}/brick-b-test1b.fits'.format(self.testDir), '_TRUTH')
+        t2 = fits.getdata('{}/brick-b-test2.fits'.format(self.testDir), '_TRUTH')
+        self.assertTrue(np.all(t1a['TRUE_OBJTYPE'] == t1b['TRUE_OBJTYPE']))
+        self.assertTrue(np.all(t1b['TRUEZ'] == t1b['TRUEZ']))
+        self.assertTrue(np.any(t1a['TRUEZ'] != t2['TRUEZ']))
+
+    #- Test that higher airmass makes noisier spectra using simspec as input
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
+    def test_quickgen_airmass_simspec(self):
+        night=self.night
+        camera = 'r0'
+        expid0 = 100
+        expid1 = 101
 
         # generate exposures of varying airmass
-        night=self.night
-        obs.new_exposure('dark',night=night,expid=100,nspec=1,airmass=1.5,seed=1)
-        simspec0 = io.findfile('simspec', night, 100)
-        fibermap0 = desispec.io.findfile('fibermap', night, 100)
-        opts0 = ['--simspec', simspec0, '--fibermap', fibermap0]
-
-        obs.new_exposure('dark',night=night,expid=101,nspec=1,airmass=1.0,seed=1)
-        simspec1 = io.findfile('simspec', night, 101)
-        fibermap1 = desispec.io.findfile('fibermap', night, 101)
-        opts1 = ['--simspec', simspec1, '--fibermap', fibermap1]
+        obs.new_exposure('dark',night=night,expid=expid0,nspec=1,airmass=1.5,seed=1)
+        simspec0 = io.findfile('simspec', night, expid0)
+        fibermap0 = desispec.io.findfile('fibermap', night, expid0)
+        obs.new_exposure('dark',night=night,expid=expid1,nspec=1,airmass=1.0,seed=1)
+        simspec1 = io.findfile('simspec', night, expid1)
+        fibermap1 = desispec.io.findfile('fibermap', night, expid1)
 
         # generate quickgen output for each airmass
-        desisim.scripts.quickgen.main(opts0)
-        desisim.scripts.quickgen.main(opts1)
+        cmd = "quickgen --simspec {} --fibermap {}".format(simspec0,fibermap0)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --simspec {} --fibermap {}".format(simspec1,fibermap1)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        cf0=desispec.io.read_frame(CFRAME100_PATH)
-        cf1=desispec.io.read_frame(CFRAME101_PATH)
+        cframe0=desispec.io.findfile("cframe",night,expid0,camera)
+        cframe1=desispec.io.findfile("cframe",night,expid1,camera)
+        cf0=desispec.io.read_frame(cframe0)
+        cf1=desispec.io.read_frame(cframe1)
         self.assertLess(np.median(cf0.ivar),np.median(cf1.ivar))
 
-    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
-    # test to see if decreased exposure time yields smaller ivar
-    def test_quickgen_exptime(self):
+    #- Test that higher airmass makes noisier spectra for bricks
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickbrick/specsim test on Travis')
+    def test_quickgen_airmass_brick(self):
+        nspec = 2
+        cmd = "quickgen --seed 1 -b test1 --airmass 1.5 --objtype DARK_MIX -n {} --outdir {}".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --seed 1 -b test2 --airmass 1.0 --objtype DARK_MIX -n {} --outdir {}".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        CFRAME100_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000100','cframe-r0-00000100.fits')
-        CFRAME101_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000101','cframe-r0-00000101.fits')
+        f1 = desispec.io.read_frame('{}/brick-b-test1.fits'.format(self.testDir))
+        f2 = desispec.io.read_frame('{}/brick-b-test2.fits'.format(self.testDir))
+        #- test1 has shorter exposure time thus larger errors thus smaller ivar
+        self.assertLess(np.median(f1.ivar), np.median(f2.ivar))
+
+    #- Test that shorter exposures make noisier spectra using simspec as input
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
+    def test_quickgen_exptime_simspec(self):
+        night = self.night
+        camera = 'r0'
+        expid0 = 100
+        expid1 = 101
 
         # generate exposures of varying exposure times
-        night=self.night
-        obs.new_exposure('dark',exptime=100,night=night,expid=100,nspec=1,seed=1)
-        simspec0 = io.findfile('simspec', night, 100)
-        fibermap0 = desispec.io.findfile('fibermap', night, 100)
-        opts0 = ['--simspec', simspec0, '--fibermap', fibermap0]
-
-        obs.new_exposure('dark',exptime=1000,night=night,expid=101,nspec=1,seed=1)
-        simspec1 = io.findfile('simspec', night, 101)
-        fibermap1 = desispec.io.findfile('fibermap', night, 101)
-        opts1 = ['--simspec', simspec1, '--fibermap', fibermap1]
+        obs.new_exposure('dark',exptime=100,night=night,expid=expid0,nspec=1,seed=1)
+        simspec0 = io.findfile('simspec', night, expid0)
+        fibermap0 = desispec.io.findfile('fibermap', night, expid0)
+        obs.new_exposure('dark',exptime=1000,night=night,expid=expid1,nspec=1,seed=1)
+        simspec1 = io.findfile('simspec', night, expid1)
+        fibermap1 = desispec.io.findfile('fibermap', night, expid1)
 
         # generate quickgen output for each exposure time
-        desisim.scripts.quickgen.main(opts0)
-        desisim.scripts.quickgen.main(opts1)
+        cmd = "quickgen --simspec {} --fibermap {}".format(simspec0,fibermap0)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --simspec {} --fibermap {}".format(simspec1,fibermap1)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        cf0=desispec.io.read_frame(CFRAME100_PATH)
-        cf1=desispec.io.read_frame(CFRAME101_PATH)
+        cframe0=desispec.io.findfile("cframe",night,expid0,camera)
+        cframe1=desispec.io.findfile("cframe",night,expid1,camera)
+        cf0=desispec.io.read_frame(cframe0)
+        cf1=desispec.io.read_frame(cframe1)
         self.assertLess(np.median(cf0.ivar),np.median(cf1.ivar))
 
-    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
-    # test to see if full moon yields smaller ivar than new moon
-    def test_quickgen_moonphase(self):
+    #- Test that shorter exposures make noisier spectra for bricks
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickbrick/specsim test on Travis')
+    def test_quickgen_exptime_brick(self):
+        cmd = "quickgen --seed 1 -b test1 --exptime 100 --objtype DARK_MIX -n 1 --outdir {}".format(self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --seed 1 -b test2 --exptime 1000 --objtype DARK_MIX -n 1 --outdir {}".format(self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        CFRAME100_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000100','cframe-r0-00000100.fits')
-        CFRAME101_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000101','cframe-r0-00000101.fits')
+        f1 = desispec.io.read_frame('{}/brick-b-test1.fits'.format(self.testDir))
+        f2 = desispec.io.read_frame('{}/brick-b-test2.fits'.format(self.testDir))
+        #- test1 has shorter exposure time thus larger errors thus smaller ivar
+        self.assertLess(np.median(f1.ivar), np.median(f2.ivar))
+
+    #- Test that brighter moon makes noisier spectra using simspec as input
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
+    def test_quickgen_moon_simspec(self):
+        night = self.night
+        camera = 'r0'
+        expid0 = 100
+        expid1 = 101
 
         # generate exposures
-        night=self.night
-        obs.new_exposure('bgs',night=night,expid=100,nspec=1,seed=1)
-        simspec0 = io.findfile('simspec', night, 100)
-        fibermap0 = desispec.io.findfile('fibermap', night, 100)
-        opts0 = ['--simspec', simspec0, '--fibermap', fibermap0, '--moon-phase', 0]
-
-        obs.new_exposure('bgs',night=night,expid=101,nspec=1,seed=1)
-        simspec1 = io.findfile('simspec', night, 101)
-        fibermap1 = desispec.io.findfile('fibermap', night, 101)
-        opts1 = ['--simspec', simspec1, '--fibermap', fibermap1, '--moon-phase', 1]
+        obs.new_exposure('bgs',night=night,expid=expid0,nspec=1,seed=1)
+        simspec0 = io.findfile('simspec', night, expid0)
+        fibermap0 = desispec.io.findfile('fibermap', night, expid0)
+        obs.new_exposure('bgs',night=night,expid=expid1,nspec=1,seed=1)
+        simspec1 = io.findfile('simspec', night, expid1)
+        fibermap1 = desispec.io.findfile('fibermap', night, expid1)
 
         # generate quickgen output for each moon phase
-        desisim.scripts.quickgen.main(opts0)
-        desisim.scripts.quickgen.main(opts1)
+        cmd = "quickgen --simspec {} --fibermap {} --moon-phase 0.1".format(simspec0,fibermap0)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --simspec {} --fibermap {} --moon-phase 0.9".format(simspec1,fibermap1)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        cf0=desispec.io.read_frame(CFRAME100_PATH)
-        cf1=desispec.io.read_frame(CFRAME101_PATH)
+        cframe0=desispec.io.findfile("cframe",night,expid0,camera)
+        cframe1=desispec.io.findfile("cframe",night,expid1,camera)
+        cf0=desispec.io.read_frame(cframe0)
+        cf1=desispec.io.read_frame(cframe1)
         self.assertLess(np.median(cf0.ivar),np.median(cf1.ivar))
 
-    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
-    # test to see if moon angle of 0 yeilds smaller ivar than a moon angle of 180
-    def test_quickgen_moonangle(self):
+    #- Test that brighter moon makes noisier spectra for bricks
+    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickbrick/specsim test on Travis')
+    def test_quickgen_moon_brick(self):
+        nspec = 2
+        cmd = "quickgen --seed 1 -b brightmoon --objtype BRIGHT_MIX -n {} --outdir {} --moon-phase 0.1".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
+        cmd = "quickgen --seed 1 -b crescentmoon --objtype BRIGHT_MIX -n {} --outdir {} --moon-phase 0.9".format(nspec, self.testDir)
+        quickgen.main(quickgen.parse(cmd.split()[1:]))
 
-        CFRAME100_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000100','cframe-r0-00000100.fits')
-        CFRAME101_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000101','cframe-r0-00000101.fits')
-
-        # generate exposures
-        night=self.night
-        obs.new_exposure('bgs',night=night,expid=100,nspec=1,seed=1)
-        simspec0 = io.findfile('simspec', night, 100)
-        fibermap0 = desispec.io.findfile('fibermap', night, 100)
-        opts0 = ['--simspec', simspec0, '--fibermap', fibermap0, '--moon-angle', 0]
-
-        obs.new_exposure('bgs',night=night,expid=101,nspec=1,seed=1)
-        simspec1 = io.findfile('simspec', night, 101)
-        fibermap1 = desispec.io.findfile('fibermap', night, 101)
-        opts1 = ['--simspec', simspec1, '--fibermap', fibermap1, '--moon-angle', 180]
-
-        # generate quickgen output for each moon angle
-        desisim.scripts.quickgen.main(opts0)
-        desisim.scripts.quickgen.main(opts1)
-
-        cf0=desispec.io.read_frame(CFRAME100_PATH)
-        cf1=desispec.io.read_frame(CFRAME101_PATH)
-        self.assertLess(np.median(cf0.ivar),np.median(cf1.ivar))
-
-    ### @unittest.skipIf('TRAVIS_JOB_ID' in os.environ, 'Skipping memory hungry quickgen/specsim test on Travis')
-    # test to see if moon zenith angle of 0 yeilds smaller ivar than moon zenith angle of 90
-    def test_quickgen_moonzenith(self):
-
-        CFRAME100_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000100','cframe-r0-00000100.fits')
-        CFRAME101_PATH = os.path.join(os.environ['HOME'],'desi_test_io','spectro','sim','test-quickgen','test-quickgen','exposures','20150105','00000101','cframe-r0-00000101.fits')
-
-        # generate exposures
-        night=self.night
-        obs.new_exposure('bgs',night=night,expid=100,nspec=1,seed=1)
-        simspec0 = io.findfile('simspec', night, 100)
-        fibermap0 = desispec.io.findfile('fibermap', night, 100)
-        opts0 = ['--simspec', simspec0, '--fibermap', fibermap0, '--moon-zenith', 0]
-
-        obs.new_exposure('bgs',night=night,expid=101,nspec=1,seed=1)
-        simspec1 = io.findfile('simspec', night, 101)
-        fibermap1 = desispec.io.findfile('fibermap', night, 101)
-        opts1 = ['--simspec', simspec1, '--fibermap', fibermap1, '--moon-zenith', 90]
-
-        # generate quickgen output for each moon angle
-        desisim.scripts.quickgen.main(opts0)
-        desisim.scripts.quickgen.main(opts1)
-
-        cf0=desispec.io.read_frame(CFRAME100_PATH)
-        cf1=desispec.io.read_frame(CFRAME101_PATH)
-        self.assertLess(np.median(cf0.ivar),np.median(cf1.ivar))
+        brickfile = '{}/brick-b-brightmoon.fits'.format(self.testDir)
+        brick1 = desispec.io.read_frame(brickfile)
+        brickfile = '{}/brick-b-crescentmoon.fits'.format(self.testDir)
+        brick2 = desispec.io.read_frame(brickfile)
+        #- brick1 has more moonlight thus larger errors thus smaller ivar
+        self.assertLess(np.median(brick1.ivar), np.median(brick2.ivar))
 
 #- This runs all test* functions in any TestCase class in this file
 if __name__ == '__main__':
